[rtl,LSU] Update LSU to allow capability accesses
Capability accesses are handled similarly to misaligned accesses in the
finite state machine within the load store unit. CHERI does not allow
for misaligned capability-wide accesses so all capability accesses will
always require 2 requests.

The memory interface has been widened by 1 bit to allow transferring
capability tags along with the data.

At the moment nothing has been done about misaligned capability accesses
so behaviour is untested in this case.

+++ dv/testrig/dii_toplevel_sim.cpp
@@ -44,7 +44,7 @@ struct Mem_Access {
     std::uint32_t addr;
     bool          write;
     std::uint8_t  be;
-    std::uint32_t data;
+    std::uint64_t data;
 };
 
 RVFI_DII_Execution_Packet readRVFI(Vibex_top_sram *top, bool signExtend);
